pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfylowda"
version = "0.0.2-alpha"
description = "Productionizing ComfyUI Workflows."
authors = [{name = "AYF", email = "realazthat@gmail.com"}]
license = {file = "LICENSE.md"}
readme = "README.md"
keywords = [
  "pipeline",
  "scheduler",
  "stable-diffusion",
  "generative-ai",
  "comfyui"
]
classifiers = [
  "Operating System :: OS Independent",
  "License :: OSI Approved :: MIT License",
  'Programming Language :: Python :: 3',
  'Programming Language :: Python :: 3.11'
]
dependencies = [
  "aiofiles >=23,<24",
  "aiohttp >=3,<4",
  "aioshutil >=1,<2",
  "aiosignal >=1,<2",
  "annotated-types <1",
  "anyio >=4,<5",
  "arrow >=1,<2",
  "attrs >=23,<24",
  "comfy_catapult @ git+https://github.com/realazthat/comfy-catapult.git@144c4bb236ab3020952f7ddc152d1c41a0e002bf",
  "deepdiff >6,<7",
  "fastapi==0.110.0",
  "fqdn >=1,<2",
  "frozenlist >=1,<2",
  "fsspec==2024.2.0",
  "h11 <1",
  "h2 >=4,<5",
  "hpack >=4,<5",
  "httpx==0.27.0",
  "hypercorn==0.16.0",
  "hyperframe >=6,<7",
  "idna >=3,<4",
  "isoduration >=20,<21",
  "jsonpointer >1.13",
  "jsonschema >=4,<5",
  "jsonschema-specifications==2023.12.1",
  "markdown-it-py >=3,<4",
  "mdurl >=0.1.2,<1",
  "multidict >=6,<7",
  "pydantic >=2,<3",
  "pydantic_core >=2,<3",
  "pydash >=7,<8",
  "Pygments >=2,<3",
  "python-datauri >=2,<3",
  "python-dateutil >=2,<3",
  "python-slugify >=8,<9",
  "PyYAML >=6,<7",
  "referencing >=0.1.0,<1",
  "requests >=2,<3",
  "rfc3339-validator <1",
  "rfc3986-validator >=0.1,<1",
  "rich >=13,<14",
  "rich-argparse >=1,<2",
  "rpds-py >=0.18,<1",
  "six >=1,<2",
  "sniffio >=1,<2",
  "StrEnum >=0.4,<1",
  "text-unidecode >=1,<2",
  "toml <1",
  "types-python-dateutil >=2,<3",
  "typing_extensions >=4,<5",
  "uri_template >=1,<2",
  "webcolors >=1.11,<2",
  "websockets >=12,<13",
  "yarl >=1,<2"
]

[project.optional-dependencies]
# Set of dependencies that are known to work. To add a new dependency here, add
# a version range to the `projecct.dependencies` section above, and then run
# `EXTRA=prod bash scripts/pin-extra-reqs.sh`. Optionally, delete all the
# contents to repin all the dependencies (probably a good idea to do once in a
# while).
prod = [
  "aiofiles==23.2.1",
  "aiohttp==3.9.3",
  "aioshutil==1.3",
  "aiosignal==1.3.1",
  "annotated-types==0.6.0",
  "anyio==4.3.0",
  "arrow==1.3.0",
  "attrs==23.2.0",
  "certifi==2024.2.2",
  "charset-normalizer==3.3.2",
  "comfy-catapult @ git+https://github.com/realazthat/comfy-catapult.git@144c4bb236ab3020952f7ddc152d1c41a0e002bf",
  "deepdiff==6.7.1",
  "fastapi==0.110.0",
  "fqdn==1.5.1",
  "frozenlist==1.4.1",
  "fsspec==2024.2.0",
  "h11==0.14.0",
  "h2==4.1.0",
  "hpack==4.0.0",
  "httpcore==1.0.4",
  "httpx==0.27.0",
  "hypercorn==0.16.0",
  "hyperframe==6.0.1",
  "idna==3.6",
  "isoduration==20.11.0",
  "jsonpointer==2.4",
  "jsonschema==4.21.1",
  "jsonschema-specifications==2023.12.1",
  "markdown-it-py==3.0.0",
  "mdurl==0.1.2",
  "multidict==6.0.5",
  "ordered-set==4.1.0",
  "priority==2.0.0",
  "pydantic==2.6.4",
  "pydantic-core==2.16.3",
  "pydash==7.0.7",
  "pygments==2.17.2",
  "python-datauri==2.1.0",
  "python-dateutil==2.9.0.post0",
  "python-slugify==8.0.4",
  "pyyaml==6.0.1",
  "referencing==0.34.0",
  "requests==2.31.0",
  "rfc3339-validator==0.1.4",
  "rfc3986-validator==0.1.1",
  "rich==13.7.1",
  "rich-argparse==1.4.0",
  "rpds-py==0.18.0",
  "six==1.16.0",
  "sniffio==1.3.1",
  "starlette==0.36.3",
  "strenum==0.4.15",
  "text-unidecode==1.3",
  "toml==0.10.2",
  "toml-sort==0.23.1",
  "tomlkit==0.12.4",
  "types-python-dateutil==2.9.0.20240316",
  "typing-extensions==4.10.0",
  "uri-template==1.3.0",
  "urllib3==2.2.1",
  "webcolors==1.13",
  "websockets==12.0",
  "wsproto==1.2.0",
  "yarl==1.9.4"
]
# Set of dev dependencies, all pinned, so that they are known to work. To add a
# new dependency here, add the unpinned package name here, and then run
# `EXTRA=dev bash scripts/pin-extra-reqs.sh`.
dev = [
  "aiofiles==23.2.1",
  "aiohttp==3.9.3",
  "aioshutil==1.3",
  "aiosignal==1.3.1",
  "annotated-types==0.6.0",
  "anyio==4.3.0",
  "arrow==1.3.0",
  "attrs==23.2.0",
  "autoflake==2.3.1",
  "certifi==2024.2.2",
  "cfgv==3.4.0",
  "charset-normalizer==3.3.2",
  "colorama==0.4.6",
  "comfy-catapult @ git+https://github.com/realazthat/comfy-catapult.git@144c4bb236ab3020952f7ddc152d1c41a0e002bf",
  "deepdiff==6.7.1",
  "defusedxml==0.7.1",
  "distlib==0.3.8",
  "fastapi==0.110.0",
  "filelock==3.13.3",
  "fqdn==1.5.1",
  "frozenlist==1.4.1",
  "fsspec==2024.2.0",
  "h11==0.14.0",
  "h2==4.1.0",
  "hpack==4.0.0",
  "httpcore==1.0.4",
  "httpx==0.27.0",
  "hypercorn==0.16.0",
  "hyperframe==6.0.1",
  "identify==2.5.35",
  "idna==3.6",
  "importlib-metadata==7.1.0",
  "isoduration==20.11.0",
  "isort==5.13.2",
  "jinja2==3.1.3",
  "jsonpointer==2.4",
  "jsonschema==4.21.1",
  "jsonschema-specifications==2023.12.1",
  "markdown-it-py==3.0.0",
  "markupsafe==2.1.5",
  "mdurl==0.1.2",
  "multidict==6.0.5",
  "nodeenv==1.8.0",
  "ordered-set==4.1.0",
  "pexpect==4.9.0",
  "platformdirs==4.2.0",
  "pre-commit==3.7.0",
  "priority==2.0.0",
  "ptyprocess==0.7.0",
  "pydantic==2.6.4",
  "pydantic-core==2.16.3",
  "pydash==7.0.7",
  "pyflakes==3.2.0",
  "pygments==2.17.2",
  "python-datauri==2.1.0",
  "python-dateutil==2.9.0.post0",
  "python-slugify==8.0.4",
  "pyyaml==6.0.1",
  "referencing==0.34.0",
  "requests==2.31.0",
  "rfc3339-validator==0.1.4",
  "rfc3986-validator==0.1.1",
  "rich==13.7.1",
  "rich-argparse==1.4.0",
  "rpds-py==0.18.0",
  "six==1.16.0",
  "sniffio==1.3.1",
  "snipinator==1.0.9",
  "starlette==0.36.3",
  "strenum==0.4.15",
  "text-unidecode==1.3",
  "toml==0.10.2",
  "toml-sort==0.23.1",
  "tomli==2.0.1",
  "tomlkit==0.12.4",
  "types-colorama==0.4.15.20240311",
  "types-python-dateutil==2.9.0.20240316",
  "typing-extensions==4.10.0",
  "uri-template==1.3.0",
  "urllib3==2.2.1",
  "virtualenv==20.25.1",
  "webcolors==1.13",
  "websockets==12.0",
  "wsproto==1.2.0",
  "yapf==0.40.2",
  "yarl==1.9.4",
  "zipp==3.18.1"
]

[project.urls]
Homepage = "https://github.com/realazthat/comfylowda"
Documentation = "https://github.com/realazthat/comfylowda"
Repository = "https://github.com/realazthat/comfylowda"
